Renamed granule example publishers
